Shifted location of logger.info message
Output of the logger.info message is now conditional on `np.linalg.norm(np.cross(vector, ref_vector))` being `0`. The test for zero is located before the try-except. The try-except is now conditional on the `np.linalg.norm(np.cross(vector, ref_vector))` not being `0`. If they are already aligned and the try-except doesn't run will this create issues with `x`?